python-backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
pyjwt>=2.0.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
pydantic>=2.0.0
email-validator>=1.0.0
